requirements.txt
asgiref==3.9.1
Django==5.2.5
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
python-dotenv==1.0.1
